skymarket/requirements.txt
asgiref==3.7.2
build==1.0.3
CacheControl==0.13.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cleo==2.1.0
crashtest==0.4.1
cryptography==41.0.4
defusedxml==0.8.0rc2
distlib==0.3.7
Django==5.0
django-cors-headers==4.3.1
django-filter==23.5
django-restframework==0.0.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-yasg==1.21.7
dulwich==0.21.6
fastjsonschema==2.18.1
filelock==3.12.2
flake8==6.1.0
idna==3.4
importlib-metadata==6.8.0
inflection==0.5.1
installer==0.7.0
itypes==1.2.0
jaraco.classes==3.3.0
jeepney==0.8.0
keyring==24.2.0
mccabe==0.7.0
more-itertools==10.1.0
msgpack==1.0.7
mypy-extensions==0.4.3
oauthlib==3.2.2
packaging==23.2
pexpect==4.8.0
phonenumbers==8.12.41
Pillow==10.1.0
pipenv==2023.11.15
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
pyparsing==3.0.7
pyproject_hooks==1.0.0
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0.1
rapidfuzz==3.5.2
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.8
SecretStorage==3.3.3
shellingham==1.5.3
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.12.1
trove-classifiers==2023.9.19
uritemplate==4.1.1
urllib3==2.0.6
virtualenv==20.25.0
zipp==3.17.0
